refactor(data-model-schema): no more `schema-utils`
The three things left in the utilities module go where they belong, and the
module itself goes away.

`internSchemaPairAsKey()` moves to `schema-intern.ts`, where its one dependency
-- `internSchemaAsTaggedHashString()` -- already lived, so the rewriting module
drops that import entirely. `isNontrivialSchema()` becomes a leaf module of its
own, with nothing but two type imports. What remains is
`schemaWithProperties()` and `schemaWithoutProperties()`, which is schema
rewriting, so the file is named `schema-rewrite.ts` -- a name that describes
its contents rather than labelling a residue.

Every module in the package now names one thing, and the graph stays a DAG.
